refactor: move version to generic util
