+++ packages/core/src/storage/migration/updates.ts
@@ -1,6 +1,6 @@
 import type { Agent } from '../../agent/Agent'
+import type { VersionString } from '../../utils/version'
 import type { V0_1ToV0_2UpdateConfig } from './updates/0.1-0.2'
-import type { VersionString } from './version'
 
 import { updateV0_1ToV0_2 } from './updates/0.1-0.2'
