+++ packages/core/src/utils/__tests__/version.test.ts
