+++ packages/core/src/utils/version.ts
